pseudo gen now according to shots

+++ news_sum/nsum_training.py
@@ -385,7 +385,7 @@ def TrainOneBatch(model, optimizer, scheduler, data, loss_fun, args):
 
     with torch.set_grad_enabled(True):
         embedding, score = model(frames.half())
-        loss = loss_fun(score.view(-1), scores)
+        loss = loss_fun(score.view(-1).half(), scores.half())
         loss = loss.half()
         print(f'Predicted Scores: {score.view(-1)[::10]}')
         print(f'GT Scores: {scores[::10]}')
